+++ .github/bin/release-plan/commit.mjs
@@ -40,3 +40,41 @@ export async function commitAfterPackageRelease(newVersion, packageDirectory, pa
 		});
 	});
 }
+
+export async function commitAfterDependencyUpdates() {
+	await new Promise((resolve, reject) => {
+		const commitCmd = spawn(
+			'git',
+			[
+				'commit',
+				'-am',
+				`set dependencies to newly released versions`
+			]
+		);
+
+		let stdoutBuffer = '';
+		let stderrBuffer = '';
+
+		commitCmd.stdout.on('data', (data) => {
+			stdoutBuffer += data.toString();
+		});
+
+		commitCmd.stderr.on('data', (data) => {
+			stderrBuffer += data.toString();
+		});
+
+		commitCmd.on('close', (code) => {
+			if (0 !== code) {
+				if (stderrBuffer) {
+					reject(new Error(`'git commit -am' exited with code ${code} and error message: ${stderrBuffer}`));
+					return;
+				}
+
+				reject(new Error(`'git commit -am' exited with code ${code}`));
+				return;
+			}
+
+			resolve(stdoutBuffer);
+		});
+	});
+}