+++ packages/media-query-list-parser/test/cases/mf-plain/0009.mjs
@@ -1,0 +1,13 @@
+import assert from 'assert';
+import { runTest } from '../../util/run-test.mjs';
+
+runTest(
+	'(min-width: env(safe-area-inset-top))',
+	'mf-plain/0009',
+	(actual, expected) => {
+		assert.deepStrictEqual(
+			actual,
+			expected,
+		);
+	},
+);