+++ plugins/postcss-is-pseudo-class/src/split-selectors/compound-selector-order.ts
@@ -1,11 +1,11 @@
 import parser from 'postcss-selector-parser';
 
 export function sortCompoundSelectorsInsideComplexSelector(node) {
-	if (!node || !node.nodes) {
+	if (!node || !node.nodes || node.nodes.length === 1) {
 		return;
 	}
 
-	const compoundSelectors = [];
+	const compoundSelectors: Array<Array<parser.Node>> = [];
 	let currentCompoundSelector = [];
 	for (let i = 0; i < node.nodes.length; i++) {
 		if (node.nodes[i].type === 'combinator') {
@@ -39,6 +39,7 @@ export function sortCompoundSelectorsInsideComplexSelector(node) {
 	const sortedCompoundSelectors = [];
 	for (let i = 0; i < compoundSelectors.length; i++) {
 		const compoundSelector = compoundSelectors[i];
+
 		compoundSelector.sort((a, b) => {
 			if (a.type === 'selector' && b.type === 'selector' && a.nodes.length && b.nodes.length) {
 				return selectorTypeOrder(a.nodes[0], a.nodes[0].type) - selectorTypeOrder(b.nodes[0], b.nodes[0].type);
@@ -55,14 +56,46 @@ export function sortCompoundSelectorsInsideComplexSelector(node) {
 			return selectorTypeOrder(a, a.type) - selectorTypeOrder(b, b.type);
 		});
 
+		const compoundSelectorChildTypes = new Set(compoundSelector.map(x => x.type));
+		const skipUniversals = compoundSelectorChildTypes.has('universal') && (
+			compoundSelectorChildTypes.has('tag') ||
+			compoundSelectorChildTypes.has('attribute') ||
+			compoundSelectorChildTypes.has('class') ||
+			compoundSelectorChildTypes.has('id') ||
+			compoundSelectorChildTypes.has('pseudo'));
+
 		for (let j = 0; j < compoundSelector.length; j++) {
+			if (compoundSelector[j].type === 'universal' && skipUniversals) {
+				compoundSelector[j].remove();
+				continue;
+			}
+
 			sortedCompoundSelectors.push(compoundSelector[j]);
 		}
 	}
 
 	for (let i = sortedCompoundSelectors.length - 1; i >= 0; i--) {
+		const previous = sortedCompoundSelectors[i - 1];
+
 		sortedCompoundSelectors[i].remove();
-		node.prepend(sortedCompoundSelectors[i]);
+
+		if (previous && previous.type === 'tag' && sortedCompoundSelectors[i].type === 'tag') {
+			const wrapped = parser.pseudo({
+				value: ':is',
+				nodes: [
+					parser.selector({
+						value: '',
+						nodes: [
+							sortedCompoundSelectors[i],
+						],
+					}),
+				],
+			});
+
+			node.prepend(wrapped);
+		} else {
+			node.prepend(sortedCompoundSelectors[i]);
+		}
 	}
 }