+++ plugins/postcss-custom-media/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Custom Media
 
+### Unreleased (patch)
+
+- Updated `@csstools/css-tokenizer` to `2.1.1` (patch)
+
 ### 9.1.2 (February 8, 2023)
 
 - Fixed: exception on missing media queries.