+++ plugins/postcss-custom-media/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Updated `@csstools/css-tokenizer` to `2.1.1` (patch)
 - Updated `@csstools/css-parser-algorithms` to `2.1.1` (patch)
-- Updated `@csstools/cascade-layer-name-parser` to `1.0.2` (patch)- Updated `@csstools/media-query-list-parser` to `2.0.4` (patch)
+- Updated `@csstools/cascade-layer-name-parser` to `1.0.2` (patch)
+- Updated `@csstools/media-query-list-parser` to `2.0.4` (patch)
 
 ### 9.1.2 (February 8, 2023)